Align MPLS tests to OVS implementation
In my reading there default values that an MPLS LSE should take
on MPLS push is not entirely clear in some circumstances.

This patch illustrates some discrepancies in the Ryu and Open vSwtich
implementations by moving the Ryu tests to match that of Open vSwtich.

* When pushing an MPLS LSE onto a packet which is neither IP nor MPLS
  use 64 as the TTL of the new LSE.

  Ryu currently uses 0.

* When pushing an MPLS LSE onto an IPv6 packet
  use 2 (IPv6 Explicit NULL) as the default label.

  Ryu currently uses 0 (IPv4 Explicit NULL).

Another approach, which would neatly bypass the lack of clarity in
the spec would be to perform explicit set_mpls_ttl and set_label
actions as part of the tests.

Signed-off-by: Simon Horman <horms@verge.net.au>