of14: Add OFPPropCommonExperimenter4ByteData
There are several different experimenter properties in OF1.4 which
are identical in structure. They have experimenter, exp_type
and data fields. The data field is an array of 4-byte words.

This patch adds OFPPropCommonExperimenter4ByteData which provides
a parser and serializer for such properties and can be inherited
by a subclass which implements one of the properties.

This patch also uses OFPPropCommonExperimenter4ByteData to implement
queue desc properties and table mod properties.

It is planned to use OFPPropCommonExperimenter4ByteData to implement
role properties.

Signed-off-by: Simon Horman <horms@verge.net.au>